Bumped wrp version from v2 to v3
