+++ WRPAccessControl_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-kit/kit/metrics"
 	"github.com/stretchr/testify/assert"
 	"github.com/xmidt-org/bascule"
-	"github.com/xmidt-org/wrp-go/v2"
+	"github.com/xmidt-org/wrp-go/v3"
 )
 
 func TestAuthorizeWRP(t *testing.T) {
