+++ mocks_test.go
@@ -2,9 +2,12 @@ package main
 
 import (
 	"context"
+	"net/http"
+	"net/http/httptest"
 
 	"github.com/stretchr/testify/mock"
-	"github.com/xmidt-org/wrp-go/v2"
+	"github.com/xmidt-org/wrp-go/v3"
+	"github.com/xmidt-org/wrp-go/v3/wrphttp"
 )
 
 type mockWRPAccessAuthority struct {
@@ -14,4 +17,26 @@ type mockWRPAccessAuthority struct {
 func (m *mockWRPAccessAuthority) authorizeWRP(ctx context.Context, message *wrp.Message) (bool, error) {
 	arguments := m.Called(ctx, message)
 	return arguments.Bool(0), arguments.Error(1)
+}
+
+type testWRPResponseWriter struct {
+	http.ResponseWriter
+}
+
+func (t *testWRPResponseWriter) WriteWRP(i *wrphttp.Entity) (int, error) {
+	return 0, nil
+}
+
+func (t *testWRPResponseWriter) WriteWRPBytes(_ wrp.Format, _ []byte) (int, error) {
+	return 0, nil
+}
+
+func (t *testWRPResponseWriter) WRPFormat() wrp.Format {
+	return wrp.Msgpack
+}
+
+func newTestWRPResponseWriter(w *httptest.ResponseRecorder) *testWRPResponseWriter {
+	return &testWRPResponseWriter{
+		ResponseWriter: w,
+	}
 }
